# skills/docmost.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Docmost Agent API (AgentWiki)",
    "version": "0.1.0",
    "description": "Filtered OpenAPI for external agents. Docmost is a thin API (store/retrieve/relay); the agent is the brain. All paths are under <DOCMOST_BASE_URL>/api. Responses are wrapped as { data, success, status } unless noted; read `data`."
  },
  "servers": [{ "url": "{baseUrl}/api", "variables": { "baseUrl": { "default": "https://wiki.example.com" } } }],
  "security": [{ "bearerAuth": [] }],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "description": "Personal API key (Settings > API keys)." }
    },
...
  },
  "paths": {
...
    "/dedup/analyze": {
      "post": {
        "operationId": "wiki_dedup_analyze",
        "summary": "Cluster exact-duplicate pages with a keep-oldest recommendation.",
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "spaceId": { "type": "string" } } } } } },
        "responses": { "200": { "description": "analysis", "content": { "application/json": { "schema": { "type": "object", "properties": { "scanned": { "type": "integer" }, "duplicateClusters": { "type": "integer" }, "clusters": { "type": "array", "items": { "type": "object" } } } } } } } }
      }
    },
    "/dedup/resolve": {
      "post": {
        "operationId": "wiki_dedup_resolve",
        "summary": "Soft-delete duplicate pages, keeping one.",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["keepPageId", "dropPageIds"], "properties": { "keepPageId": { "type": "string", "format": "uuid" }, "dropPageIds": { "type": "array", "items": { "type": "string", "format": "uuid" } }, "mode": { "type": "string", "enum": ["soft-delete"] } } } } } },
        "responses": { "200": { "description": "result", "content": { "application/json": { "schema": { "type": "object" } } } } }
      }
    },
...
  }
}